Rename the plugin to Grafana-dashboard plugin
