+++ plugins/grafana_infiniband_telemetry_plugin/conf/grafana-dashboard_httpd_proxy.conf
