+++ plugins/grafana_infiniband_telemetry_plugin/conf/grafana_infiniband_telemetry_plugin.cfg
@@ -7,7 +7,7 @@ url = enterprise
 port = 8000
 
 [logs-config]
-logs_file_name = /log/git.log
+logs_file_name = /log/grafana-dashboard-plugin.log
 logs_level = INFO
 log_file_max_size = 10485760
 log_file_backup_count = 5
